requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.93.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
tomli==2.0.1
typing-extensions==4.5.0
uvicorn==0.21.0
